+++ src/main/kotlin/blackjack/domain/CardPlayer.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 interface CardPlayer {
     val name: String
