+++ src/main/kotlin/blackjack/domain/Draw.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 class Draw(
     val nextCard: () -> Card,
