+++ src/main/kotlin/blackjack/domain/PlayResult.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 enum class PlayResult {
     WINS, LOSSES, DRAWS
