+++ src/test/kotlin/blackjack/view/UserInputTest.kt
@@ -1,5 +1,6 @@
-package blackjack
+package blackjack.view
 
+import blackjack.view.UserInput
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
